apply patch from Ingo van Lil explanation copied form his email see below
one of my former colleagues at university sent me a mail and reported
frequent assertion failures in my TCM2 compass module. It looks like the
errors are caused by inaccurate double/float rounding in several
integrateData() methods: All my values are normalized by the
Miro::Angle class, explicitly cast to float, and finally compared with
M_PI. This comparison performs an implicit cast back to double, and it
looks like these casts can make the value leave the allowed range.

Fixing the problem is quite simple: Explicitly casting the double
constants to float before comparing should avoid any implicit casts.
A Patch for the CVS version is attached.

+++ trunk/src/miro/CompassImpl.cpp
@@ -137,8 +137,8 @@ namespace Miro
   void 
   CompassImpl::integrateData(const CompassEventIDL& _data)
   {
-    assert(_data.heading > -M_PI &&
-	   _data.heading <= M_PI);
+    assert(_data.heading > (float)-M_PI &&
+	   _data.heading <= (float)M_PI);
 
     { // scope for guard
       Guard guard(mutex_);
